Move gameDataTypes to types folder
